Enable NarrowingCompoundAssignment error-prone check
In Java, compound assignment with wider type on the right-hand-side is a
silent truncation.

This enables error-prone check for such cases and refactors code to make
it compliant. Some places are updated based solely on the assumption
that they weren't incorrect.

+++ core/trino-main/src/main/java/io/trino/operator/scalar/SequenceFunction.java
@@ -97,7 +97,7 @@ public final class SequenceFunction
 
         BlockBuilder blockBuilder = DATE.createBlockBuilder(null, length);
 
-        int value = 0;
+        long value = 0;
         for (int i = 0; i < length; ++i) {
             DATE.writeLong(blockBuilder, DateTimeOperators.datePlusIntervalYearToMonth(start, value));
             value += step;
